Correct format

+++ src/column_definitions.py
@@ -72,7 +72,7 @@ SIMPLE = [
     ("GENE_ID_ENTREZ", "Gene ID in Entrez", Type.STRING),
     ("EXPRESSION_NORM", "", Type.DOUBLE),
     ("GFF_SCORE", "A floating point value. The type is a string because the value '.' is also acceptable.", Type.STRING),
-    ("PEAK_TAG_DENSITY", "Peak tag density", Type.DOUBLE),
+    ("PEAK_TAG_DENSITY", "Peak tag density", Type.INTEGER),
     ("Z_SCORE", "Z-Score", Type.DOUBLE)
 ]
